refactor: crew 목록 페이지 컴포넌트 폴더 구조 변경
페이지에서만 사용되는 컴포넌트 _components 폴더 내부로 이동
