+++ src/app/crews/_components/CrewFilterBar/index.tsx
@@ -1,0 +1,28 @@
+import OptionDropdown from '@/components/ui/OptionDropdown';
+import { CREW_SORT_OPTIONS } from '@/constants/crew';
+import { CrewListFilters } from '@/types';
+import CrewRegionFilter from './CrewRegionFilter';
+
+interface CrewFilterBarProps {
+  filters: CrewListFilters;
+  applyFilters: (filters: CrewListFilters) => void;
+}
+
+export default function CrewFilterBar({
+  filters,
+  applyFilters,
+}: CrewFilterBarProps) {
+  return (
+    <div className="flex w-full items-center justify-between">
+      <CrewRegionFilter
+        value={filters.city}
+        onChange={(city) => applyFilters({ ...filters, city })}
+      />
+      <OptionDropdown
+        options={CREW_SORT_OPTIONS}
+        value={filters.sort || CREW_SORT_OPTIONS[0].value}
+        onChange={(sort) => applyFilters({ ...filters, sort })}
+      />
+    </div>
+  );
+}
